Added note about xref validity in shortdesc
DITA 1.3 made it possible to use an xref element within a shortdesc element. The Style Guide was written around DITA 1.1, with some provision for 1.2. Added a note in the c_Formatting_in_shortdesc.dita topic to point out that the statement about xref having to be wrapped in a ph related to "before 1.3". However, the suggestion to avoid xref still remains solid, on the basis that shortdesc should be viewed as metadata, and xref elements are incompatible with metadata.